Revert "Organising files."
This reverts commit 390084734c001c591fc289c27ded2119cedd4999.
