+++ AVSCAN/src/main/java/com/mycompany/avscan/CheckIn.java
@@ -4,7 +4,6 @@
  */
 package com.mycompany.avscan;
 
-import Database.DatabaseUtils;
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.JTextField;
