+++ AVSCAN/src/main/java/com/mycompany/avscan/DatabaseUtils.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package Database;
+package com.mycompany.avscan;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
